:recycle: refactor : move files of user.model.response to user.exception
