+++ src/main/java/com/run_us/server/domains/user/exception/UserException.java
@@ -1,4 +1,4 @@
-package com.run_us.server.domains.user.model.response;
+package com.run_us.server.domains.user.exception;
 
 import com.run_us.server.global.exceptions.BusinessException;
 import com.run_us.server.global.exceptions.enums.CustomResponseCode;
